remove unnecessary back buttons, add logout button in officer side bar, remove in officer dashboard [Elisha]
