+++ src/pages/myevents.tsx
@@ -167,13 +167,6 @@ const MyEvents: React.FC = () => {
       <MemberSidebar />
       <main className="ml-64 p-8">
         <div className="py-2">
-          <Link
-            href="/memberpage"
-            className="flex items-center space-x-1 text-gray-600 hover:text-gray-800"
-          >
-            <ArrowBackIcon />
-            <span>Back to Dashboard</span>
-          </Link>
         </div>
 
         <div className="mt-6">
